backend/ai_engine/requirements.txt
pandas
scikit-learn
pgmpy
pyyaml
openpyxl
rpy2
matplotlib
seaborn
